working examples

+++ demos/deamons/async_orchestration.py
@@ -1,4 +1,4 @@
-def init_myc():
+def init_myc(drop=False):
     import pyArango.connection as ADB
     import blumycelium.mycelium as myc
 
@@ -16,12 +16,13 @@ def init_myc():
         name="animals_demo_mycelium"
     )
 
+    if drop:
+        #Drop all previous job information from the mycelium
+        mycelium.drop()
+    
     #Ensure the mycelium is initialized
     mycelium.init(init_db=True)
 
-    #Drop all previous job information from the mycelium
-    mycelium.drop()
-    
     return mycelium
 
 def main():
@@ -34,12 +35,14 @@ def main():
     import elves
     import os
     import json
+    import time
 
-    json_database_filename = "my_animals.json"
-    with open(json_database_filename, "w") as fi:
-        json.dump({}, fi)
+    if not os.path.exists("my_animals.json"):
+        json_database_filename = "my_animals.json"
+        with open(json_database_filename, "w") as fi:
+            json.dump({}, fi)
 
-    mycelium = init_myc()
+    mycelium = init_myc(drop=True)
 
     #Ensure elves are registered in the mycelium
     animals = elves.Animals("animals creator", mycelium) 
@@ -55,29 +58,28 @@ def main():
     #this is what we would do to compute stats in parallel 
     mean_calc = elves.Stats("calc1", mycelium)
     mean_calc.register()
-    mean_calc.set_database(json_database_filename)
     
     min_calc = elves.Stats("calc2", mycelium)
     min_calc.register()
-    min_calc.set_database(json_database_filename)
     
     max_calc = elves.Stats("calc3", mycelium)
     max_calc.register()
-    max_calc.set_database(json_database_filename)
     
     printer = elves.Formater("report printer", mycelium)
     printer.register()
 
     #instanciating the jobs in the mycelium with 
-    for nb in range(1000):
+    for nb in range(100):
+        print("Sending: %s" % nb)
         mesurement = animals.task_get_animal_data()
         store.task_save_animal_data(species=mesurement["species"], weight=mesurement["weight"])
         #print stats every 10 iterations
-        if nb % 10 ==0:
+        if nb % 5 ==0:
             mean = mean_calc.task_calculate_means()
             mins = min_calc.task_calculate_mins()
             maxs = max_calc.task_calculate_maxs()
             printer.task_print_stats(means=mean["means"], mins=mins["mins"], maxs=maxs["maxs"])
-
+        time.sleep(1)
+    
 if __name__ == '__main__':
     main()
